requirements.txt
psycopg2
Flask-JWT-Extended
Flask-WTF